requirements.txt
beautifulsoup4==4.8.1
cffi==1.12.3
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mechanize3==0.4.0
pycparser==2.19
six==1.12.0
soupsieve==1.9.4
webencodings==0.5.1
Werkzeug==0.16.0
